Created ThemeContextProvider and useTheme hook
