+++ app/layout.tsx
@@ -5,6 +5,7 @@ import ActiveSectionContextProvider from "@/context/active-section-context";
 import { Toaster } from "react-hot-toast";
 import Footer from "@/components/footer";
 import ThemeSwitch from "@/components/theme-switch";
+import ThemeContextProvider from "@/context/theme-context";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -22,13 +23,17 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
             >
                 <div className="bg-[#fbe2e3] absolute top-[-6rem] -z-10 right-[11rem] h-[31.25rem] w-[31.25rem] rounded-full blur-[10rem] sm:w-[68.75rem] dark:bg-[#946263]"></div>
                 <div className="bg-[#dbd7fb] absolute top-[-1rem] -z-10 left-[-35rem] h-[31.25rem] w-[50rem] rounded-full blur-[10rem] sm:w-[68.75rem] md:left-[-33rem] lg:left-[-28rem] xl:left-[-15rem] 2xl:left-[-5rem] dark:bg-[#676394]"></div>
-                <ActiveSectionContextProvider>
-                    <Header />
-                    {children}
-                    <Toaster position="top-center" />
-                    <Footer />
-                </ActiveSectionContextProvider>
-                <ThemeSwitch />
+
+                <ThemeContextProvider>
+                    <ActiveSectionContextProvider>
+                        <Header />
+                        {children}
+                        <Toaster position="top-center" />
+                        <Footer />
+                    </ActiveSectionContextProvider>
+
+                    <ThemeSwitch />
+                </ThemeContextProvider>
             </body>
         </html>
     );
