+++ components/theme-switch.tsx
@@ -1,38 +1,10 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import { useTheme } from "@/context/theme-context";
 import { BsMoon, BsSun } from "react-icons/bs";
 
-type Theme = "light" | "dark";
-
 export default function ThemeSwitch() {
-    const [theme, setTheme] = useState<Theme>("light");
-
-    const toggleTheme = () => {
-        if (theme === "light") {
-            setTheme("dark");
-            document.documentElement.classList.add("dark");
-            window.localStorage.setItem("theme", "dark");
-        } else {
-            setTheme("light");
-            window.localStorage.setItem("theme", "light");
-            document.documentElement.classList.remove("dark");
-        }
-    };
-
-    useEffect(() => {
-        const localTheme = window.localStorage.getItem("theme") as Theme | null;
-
-        if (localTheme) {
-            setTheme(localTheme);
-            if (localTheme === "dark") {
-                document.documentElement.classList.add("dark");
-            }
-        } else if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
-            setTheme("dark");
-            document.documentElement.classList.add("dark");
-        }
-    }, []);
+    const { theme, toggleTheme } = useTheme();
 
     return (
         <button
